Python/requirements.txt
fastapi
uvicorn
langchain==0.3.25
langchain-core==0.3.65
langchain-cohere==0.3.4
langchain-community==0.3.24
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
faiss-cpu
pypdf
groq
cohere==5.13.0
python-dotenv